+++ src/index.js
@@ -14,7 +14,7 @@ import { render } from 'react-dom';
 import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
 import App from './App';
-import store from './store';
+import store from './redux/store';
 import './index.css';
 
 render(
